Removed dependant libraries and errant hidden files
Looks like it's more trouble than it's worth to package theses timer libraries in the same folder. Also, when I originally wrote this, it was on a Mac and Mac OS had placed a bunch of hidden files in the library folder that were now messing with compilation on a Windows machine.